third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/initial-empty-document/load-event-iframe-element-expected.txt

This is a testharness.js-based test.
PASS load event fires synchronously on <iframe> element created with no src
PASS load event fires synchronously on <iframe> element created with src=''
PASS load event fires synchronously on <iframe> element created with src='about:blank'
FAIL load event fires synchronously on <iframe> element created with src='about:blank#foo' assert_equals: expected 1 but got 0
PASS load event fires synchronously on <iframe> element created with src='about:blank?foo'
Harness: the test ran to completion.
